requirements.txt
streamlit==1.38.0
mysql-connector-python==8.0.33
pandas==2.2.2
openpyxl==3.1.2
python-docx==1.1.0